+++ tests/unit/bff/v1/routers/test_img2txt.py
@@ -1,19 +1,11 @@
 import pytest
 
 
-def test_index(test_client, test_index_image):
+def test_index(test_client, test_index_text):
     with pytest.raises(ConnectionError):
         test_client.post(
-            f'/api/v1/image/index',
-            json=test_index_image,
-        )
-
-
-def test_search(test_client, test_search_text):
-    with pytest.raises(ConnectionError):
-        test_client.post(
-            f'/api/v1/image/search',
-            json=test_search_text,
+            f'/api/v1/image-to-text/index',
+            json=test_index_text,
         )
 
 
@@ -39,12 +31,4 @@ def test_search_no_query(test_client):
         test_client.post(
             f'/api/v1/image/search',
             json={},
-        )
-
-
-def test_search_both_query(test_client, test_search_both):
-    with pytest.raises(ValueError):
-        test_client.post(
-            f'/api/v1/image/search',
-            json=test_search_both,
         )
